extract merge_dicts from app into utils and add tests.
