+++ antminer_autotune/util.py
@@ -15,3 +15,13 @@ def makedir(path):
 
 def fix_json_format(bad_json):
     return bad_json.replace('}{', '},{').strip(' \0')
+
+
+def merge_dicts(*dict_args):
+    """Given any number of dicts, shallow copy and merge into a new dict,
+    precedence goes to key value pairs in latter dicts."""
+    result = {}
+    for dictionary in dict_args:
+        result.update(dictionary)
+    return result
+
